Make kj = ESC. Fix bug in macro recording.
Added a timeout while waiting for j to be pushed after k so that it wouldn't drop out of insert mode unless they were pushed rapidly together.

Added a little code to make the undo handling work right (in most cases) when kj has been used. There is still a problem when you use it during macro recording though. For now, just use escape instead if you're recording a macro.

Macro recording wasn't working right because of a call to a non-existant addon for printing output to the status bar. For now I've just removed the calls.